+++ src/components/contact/index.ts
@@ -1,0 +1,2 @@
+export { ContactInfoSidebar } from "./ContactInfoSidebar";
+export { ContactForm } from "./ContactForm";